refs #22 Move the tests timeout into a constant
Configure the service created in the unit test to use timeout value specified for tests.
Additionally lower the timeout value to 5 seconds, that should be enough in most cases.
And also reconfigure the cancellation test to fulfill the cancellation expectation in time exceeding the timeout to make sure that a network request was really cancelled.