refs #27 Setup a Gitlab CI config
Add a job for building an .xcframework from archives of macOS and iOS (device + simulator) targets and use it as an artifact of the job.
The build job is performed on release tags.

Also add two jobs for making a release on a GItlab and Github for an .xcframework built for a release tag.